Updated to version 1.02.
Mod is no longer needed to be installed in client if it is installed in server.
Client will now check if server has the mod installed and handle functions on client or server depending on that.
Changed version to 1.02 in QuickHotbarModInfo.
Armor icons no longer render on top of the inventory preview.